Refactor AgentProfile
